server/requirements.txt
flask
numpy
opencv-python
cairosvg
pyserial
